+++ stralloc.h
@@ -79,7 +79,7 @@ typedef struct malloc_block_s {
  * sv->subtype is STRING_MALLOC or STRING_SHARED, and runs significantly
  * faster.
  */
-#define COUNTED_STRLEN(x) ((svalue_strlen_size = MSTR_SIZE(x)), svalue_strlen_size != USHRT_MAX ? svalue_strlen_size : strlen((x)+USHRT_MAX)+USHRT_MAX)
+#define COUNTED_STRLEN(x) ((svalue_strlen_size = MSTR_SIZE(x)), svalue_strlen_size != USHRT_MAX ? (size_t)svalue_strlen_size : strlen((x)+USHRT_MAX)+USHRT_MAX)
 /* return the number of references to a STRING_MALLOC or STRING_SHARED 
    string */
 #define COUNTED_REF(x)    MSTR_REF(x)
